fix: change variable type

+++ src/main/java/com/daigo/springboot_handson_4/domains/ResultInfo.java
@@ -11,7 +11,7 @@ public class ResultInfo {
     private Integer Count;
     private Integer Total;
     private Integer Start;
-    private Integer Latency;
+    private Double Latency;
     private Integer Status;
     private String Description;
     private String Copyright;
